+++ CHANGELOG.md
@@ -21,6 +21,7 @@ As a minor extension, we have adopted a slightly different versioning convention
     - Added new `CardanoBlocksTransactions` signed entity type for certifying both Cardano blocks and transactions.
     - Support for computing and sending signatures for the `CardanoBlocksTransactions` signed entity type in the signer.
     - Support for verifying and aggregating signatures for the `CardanoBlocksTransactions` signed entity type in the aggregator.
+    - Support for computing proofs of inclusions for a list of Cardano blocks or transactions hashes in the aggregator; the proof includes extra metadata about the included blocks or transactions.
 
 - **UNSTABLE**:
   - Support for **DMQ consumer deduplication** to prevent processing the same message multiple times upon reconnection to the DMQ server.